+++ cloudtrail.tf
@@ -1,0 +1,70 @@
+locals {
+  trail_s3_key_prefix = "cloudtrail"
+  trail_bucket_name   = "${var.app_name_prefix}-${terraform.workspace}-trail"
+}
+################## CLOUDTRAIL ##################
+
+resource "aws_cloudtrail" "main_trail" {
+  name                          = "${var.app_name_prefix}-${terraform.workspace}-trail"
+  s3_bucket_name                = aws_s3_bucket.cloudtrail.id
+  s3_key_prefix                 = local.trail_s3_key_prefix
+  include_global_service_events = true
+  enable_log_file_validation    = true
+  enable_logging                = true
+  is_multi_region_trail         = true
+  tags                          = local.global_tags
+}
+
+
+################## S3 (Trail logs) ##################
+
+resource "aws_s3_bucket" "cloudtrail" {
+  bucket        = local.trail_bucket_name
+  force_destroy = false
+  server_side_encryption_configuration {
+    rule {
+      apply_server_side_encryption_by_default {
+        sse_algorithm = "AES256"
+      }
+    }
+  }
+
+  policy = <<POLICY
+{
+    "Version": "2012-10-17",
+    "Statement": [
+        {
+            "Sid": "AWSCloudTrailAclCheck",
+            "Effect": "Allow",
+            "Principal": {
+              "Service": "cloudtrail.amazonaws.com"
+            },
+            "Action": "s3:GetBucketAcl",
+            "Resource": "arn:aws:s3:::${local.trail_bucket_name}"
+        },
+        {
+            "Sid": "AWSCloudTrailWrite",
+            "Effect": "Allow",
+            "Principal": {
+              "Service": "cloudtrail.amazonaws.com"
+            },
+            "Action": "s3:PutObject",
+            "Resource": "arn:aws:s3:::${local.trail_bucket_name}/${local.trail_s3_key_prefix}/AWSLogs/${data.aws_caller_identity.current.account_id}/*",
+            "Condition": {
+                "StringEquals": {
+                    "s3:x-amz-acl": "bucket-owner-full-control"
+                }
+            }
+        }
+    ]
+}
+POLICY
+}
+
+resource "aws_s3_bucket_public_access_block" "cloudtrail_bucket" {
+  bucket                  = aws_s3_bucket.cloudtrail.id
+  block_public_acls       = true
+  block_public_policy     = true
+  ignore_public_acls      = true
+  restrict_public_buckets = true
+}