+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Cloudwatch log group for CodeBuild
+- Cloudtrail
+- Trail bucket (for Cloudtrail logs)
 
 ## [0.4.0] - 2021-09-21
 ### Added